refactor: add MVC para módulo de funcionários
